+++ modules/graphics/functions/plot.m
@@ -101,14 +101,65 @@ function q = completeProperties(cs, ms, ps, p)
   end
 end
 %=============================================================================
-function k = plotVector(go, x, y, lineProperties)
-  index = length(go.Children) + 1;
+function color = getColorAndUpdateIndex(go)
   colorOrder = go.ColorOrder;
-  indexmod = uint32(mod(index-1, size(colorOrder, 1)) + 1);
-  if (~any(strcmp(lineProperties, 'color')))
-    lineProperties = [lineProperties, {'MarkerEdgeColor', colorOrder(indexmod,:), 'MarkerFaceColor', colorOrder(indexmod, :)}];
+  if isempty (colorOrder)
+    color = [0 0 0];
+    return;
   end
-  k = __line__('XData', x, 'YData', y, 'Color', colorOrder(indexmod, :), lineProperties{:});
+  colorIndex = go.ColorOrderIndex;
+  nbColors = size (colorOrder, 1);
+  colorIndex = mod (colorIndex, nbColors);
+  if (colorIndex == 0)
+    colorIndex = nbColors;
+  elseif (colorIndex < 0)
+    colorIndex = 1;
+  end
+  color = colorOrder(colorIndex, :);
+  if (colorIndex >= nbColors)
+    colorIndex = colorIndex + 1;
+    colorIndex = mod (colorIndex, nbColors);
+    if (colorIndex == 0)
+      colorIndex = 1;
+    end
+    lineStyleOrderIndex = go.LineStyleOrderIndex;
+    go.LineStyleOrderIndex = lineStyleOrderIndex + 1;
+  else
+     colorIndex = colorIndex + 1;
+  end
+  go.ColorOrderIndex = colorIndex;
+end
+%=============================================================================
+function lineStyle = getLineStyleAndUpdateIndex(go)
+  lineStyleOrder = go.LineStyleOrder;
+  if isempty (lineStyleOrder)
+    lineStyle = '-';
+    return;
+  end
+  if ischar(lineStyleOrder)
+    lineStyleOrder = cellstr(lineStyleOrder);
+  end
+  style_idx = go.LineStyleOrderIndex;
+  num_styles = size(lineStyleOrder, 1);
+  style_idx = mod (style_idx, num_styles);
+  if (style_idx == 0)
+    style_idx = num_styles;
+  elseif (style_idx < 0)
+    style_idx = 1;
+  end
+  lineStyle = lineStyleOrder{style_idx};
+end
+%=============================================================================
+function k = plotVector(go, x, y, lineProperties)
+  lineStyle = getLineStyleAndUpdateIndex(go);
+  color = getColorAndUpdateIndex(go);
+  if (~any(strcmp(lineProperties, 'Color')))
+    lineProperties = [lineProperties, {'MarkerEdgeColor', color, 'MarkerFaceColor', color}];
+  end
+  if (~any(strcmp(lineProperties, 'LineStyle')))
+    lineProperties = [lineProperties, {'LineStyle', lineStyle}];
+  end  
+  k = __line__('XData', x, 'YData', y, 'Color', color, lineProperties{:});
 end
 %=============================================================================
 function h = plot_XY(X, Y, go, lineProperties)